parsing-api/requirements.txt
# FastAPI framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# PDF processing (free, open-source)
PyMuPDF==1.23.22
Pillow==10.2.0

# Database & storage
supabase==2.3.0
python-dotenv==1.0.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
